Move fetchSiteInfo to the repository
