+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/sitepicker/SitePickerRepository.kt
@@ -7,6 +7,7 @@ import kotlinx.coroutines.suspendCancellableCoroutine
 import kotlinx.coroutines.withContext
 import org.greenrobot.eventbus.Subscribe
 import org.greenrobot.eventbus.ThreadMode
+import org.greenrobot.eventbus.ThreadMode.MAIN
 import org.wordpress.android.fluxc.Dispatcher
 import org.wordpress.android.fluxc.action.AccountAction
 import org.wordpress.android.fluxc.generated.AccountActionBuilder
@@ -15,8 +16,13 @@ import org.wordpress.android.fluxc.model.SiteModel
 import org.wordpress.android.fluxc.store.AccountStore
 import org.wordpress.android.fluxc.store.AccountStore.OnAccountChanged
 import org.wordpress.android.fluxc.store.SiteStore
+import org.wordpress.android.fluxc.store.SiteStore.ConnectSiteInfoPayload
+import org.wordpress.android.fluxc.store.SiteStore.OnConnectSiteInfoChecked
+import org.wordpress.android.fluxc.store.SiteStore.SiteErrorType
 import org.wordpress.android.fluxc.store.WooCommerceStore
 import org.wordpress.android.login.util.SiteUtils
+import org.wordpress.android.util.AppLog
+import org.wordpress.android.util.AppLog.T.API
 import javax.inject.Inject
 import kotlin.coroutines.resume
 
@@ -86,4 +92,32 @@ class SitePickerRepository @Inject constructor(
             dispatcher.unregister(listener)
         }
     }
+
+    suspend fun fetchSiteInfo(siteAddress: String) =
+        suspendCancellableCoroutine<Result<ConnectSiteInfoPayload>> { continuation ->
+            val listener = object : Any() {
+                @Subscribe(threadMode = MAIN)
+                fun onFetchedConnectSiteInfo(event: OnConnectSiteInfoChecked) {
+                    dispatcher.unregister(this)
+                    if (!continuation.isActive) return
+
+                    if (event.isError) {
+                        AppLog.e(API, "onFetchedConnectSiteInfo has error: " + event.error.message)
+                        continuation.resume(
+                            Result.failure(FetchSiteInfoException(event.error.type, event.error.message))
+                        )
+                    } else {
+                        continuation.resume(Result.success(event.info))
+                    }
+                }
+            }
+            dispatcher.register(listener)
+            dispatcher.dispatch(SiteActionBuilder.newFetchConnectSiteInfoAction(siteAddress))
+
+            continuation.invokeOnCancellation {
+                dispatcher.unregister(listener)
+            }
+        }
+
+    class FetchSiteInfoException(val type: SiteErrorType, message: String?) : Exception(message)
 }
